Split stripe callbacks in events file
